linh code facade
