+++ src/graphControl/+state/facade.js
@@ -1,1 +1,31 @@
-export default function useFacade() {}
+import {
+  selectErrorMessager,
+  selectLoading,
+  selectLoadSucess,
+} from "./selector";
+import { useSelector, useDispatch } from "react-redux";
+import { useEffect } from "react";
+import { getFromRefeshData } from "./effect";
+let isInitial = true;
+export default function useFacade() {
+  const data = useSelector(selectLoadSucess);
+  const waitData = useSelector(selectLoading);
+  const error = useSelector(selectErrorMessager);
+  const dispatch = useDispatch();
+  const add = "";
+  useEffect(() => {
+    if (isInitial) {
+      isInitial = false;
+      dispatch(getFromRefeshData());
+    }
+  }, [dispatch]);
+
+  return {
+    add,
+    data: data,
+    waitData,
+    error,
+    // loadData: () => getFromRefeshData(),
+    // addData: (title) => dispatch(createData({ title })),
+  };
+}
