+++ src/services/graphServices.js
@@ -7,3 +7,14 @@ export async function getGraph() {
   });
   return await response.json();
 }
+
+// export async function createGraph(charts) {
+//   const response = await fetch("http://localhost:3001/charts", {
+//     method: "POST",
+//     headers: {
+//       "Content-Type": "application/json",
+//     },
+//     body: JSON.stringify(charts),
+//   });
+//   return await response.json();
+// }
